refactor(Coin): 사용자 금액 계산 메서드 refactoring
